fix 修改属性3

+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/controller/admin/purchase/vo/in/ErpPurchaseInBaseRespVO.java
@@ -130,7 +130,7 @@ public class ErpPurchaseInBaseRespVO extends ErpPurchaseBaseRespVO {
         private int srcNo;
 
         @Schema(description = "价税合计")
-        private double allAmount;
+        private BigDecimal allAmount;
     }
 
 }
